fix pages for updated software and add parital tip
