+++ e2e-testing/automation/pd_pages/__init__.py
@@ -9,6 +9,7 @@ from .landing_page import LandingPage
 from .magnetic_module_step_form_page import AddMagneticModule
 from .mix_step_form import MixStepForm
 from .module_config_page import ModuleConfigPage
+from .partial_96ch import Partial96chScenario
 from .pipette_modal import PipetteModal
 from .protocol_editor_page import ProtocolEditorPage
 from .settings_page import SettingsPage
@@ -24,6 +25,7 @@ __all__ = [
     "LandingPage",
     "MixStepForm",
     "ModuleConfigPage",
+    "Partial96chScenario",
     "PipetteModal",
     "ProtocolEditorPage",
     "SettingsPage",
