+++ e2e-testing/tests/pd/test_pd_96ch_partial_row_column.py
@@ -1,0 +1,90 @@
+"""E2E: 96-channel row and column partial tip strategies (cascade-safe edges)."""
+
+from pathlib import Path
+
+import pytest
+from playwright.sync_api import Page
+
+from automation.pd_pages import ProtocolEditorPage, TransferPage
+from automation.pd_pages.partial_96ch import (
+    COLUMN_NOZZLE_SAMPLE_STEP_INDICES,
+    PROTOCOL_PATH,
+    ROW_NOZZLE_SAMPLE_STEP_INDICES,
+    column_nozzle_scenarios,
+    move_1000ul_tiprack_away_from_stacker,
+    print_suite_plan,
+    row_nozzle_scenarios,
+    run_partial_suite,
+)
+from utility import import_protocol_and_open_editor
+
+_ROW_LIQUID_CLASS_CONTINUE_REASON = (
+    "96ch single-row transfers hang on liquid-class Continue in this fixture "
+    "(column/single/full-rack suites pass). Cascade tip map is in place; "
+    "re-enable when row + liquid-class Continue is stable in headed runs."
+)
+
+
+@pytest.mark.pdE2E
+@pytest.mark.slow
+@pytest.mark.timeout(600)
+@pytest.mark.xfail(reason=_ROW_LIQUID_CLASS_CONTINUE_REASON, strict=False)
+def test_pd_96ch_partial_row_nozzle_tip_strategies(page: Page, pd_exports_dir: Path) -> None:
+    """Suite 2: Move + row-nozzle transfers on cascade-safe opposite tip rows."""
+    scenarios = row_nozzle_scenarios()
+    print_suite_plan(
+        "96ch row-nozzle tip strategies (A1→tip H / H1→tip A on populated rack)",
+        [
+            "Gripper Move tiprack B3→B2 (clear A4 stacker collision)",
+            *[s.label for s in scenarios],
+            f"Assert timeline sample {ROW_NOZZLE_SAMPLE_STEP_INDICES} + export 96ch_partial_row_nozzle.py",
+        ],
+    )
+
+    import_protocol_and_open_editor(page, PROTOCOL_PATH, migration=True)
+    editor = ProtocolEditorPage(page)
+    transfer = TransferPage(page)
+
+    move_1000ul_tiprack_away_from_stacker(editor)
+    run_partial_suite(
+        page,
+        editor,
+        transfer,
+        scenarios,
+        exports_dir=pd_exports_dir,
+        export_filename="96ch_partial_row_nozzle.py",
+        sample_indices=ROW_NOZZLE_SAMPLE_STEP_INDICES,
+        min_steps=1 + len(scenarios),
+    )
+
+
+@pytest.mark.pdE2E
+@pytest.mark.slow
+@pytest.mark.timeout(600)
+def test_pd_96ch_partial_column_nozzle_tip_strategies(page: Page, pd_exports_dir: Path) -> None:
+    """Suite 3: Move + column-nozzle transfers on cascade-safe opposite tip columns."""
+    scenarios = column_nozzle_scenarios()
+    print_suite_plan(
+        "96ch column-nozzle tip strategies (auto col 12 → manual A12 reuse after return-tip)",
+        [
+            "Gripper Move tiprack B3→B2 (clear A4 stacker collision)",
+            *[s.label for s in scenarios],
+            f"Assert timeline sample {COLUMN_NOZZLE_SAMPLE_STEP_INDICES} + export 96ch_partial_column_nozzle.py",
+        ],
+    )
+
+    import_protocol_and_open_editor(page, PROTOCOL_PATH, migration=True)
+    editor = ProtocolEditorPage(page)
+    transfer = TransferPage(page)
+
+    move_1000ul_tiprack_away_from_stacker(editor)
+    run_partial_suite(
+        page,
+        editor,
+        transfer,
+        scenarios,
+        exports_dir=pd_exports_dir,
+        export_filename="96ch_partial_column_nozzle.py",
+        sample_indices=COLUMN_NOZZLE_SAMPLE_STEP_INDICES,
+        min_steps=1 + len(scenarios),
+    )
